Removed unnecessary firing assert
An assert from before we had unexpected mismatch handling was left in the test
file and was incorrectly firing, removed.

This wasn't noticed in the previous commit because the log, which is stdout only
is clean, but the assert was going to stderr and it's at the end of this test so
it didn't affect other tests. Fixed the batch file to put both in the log

Files:

test/test_result_operand_types.py
- Removed assert

Tests:
Ran
    test all
Tests pass the same as before
    testing all_result_operand_types.c
    727 12.5055905187 lines/sec
    Ran 726 tests, found 0 unexpected mismatches
    testing assigns.c
    29 33.4101370742 lines/sec
    Ran 3 tests, found 0 unexpected mismatches
    testing constants.c
    45 73.1707305727 lines/sec
    Ran 3 tests, found 0 unexpected mismatches
    testing locals.c
    12 15.6657956815 lines/sec
    Ran 1 tests, found 0 unexpected mismatches
    testing ops.c
    28 40.2298793246 lines/sec
    Ran 6 tests, found 0 unexpected mismatches
    testing scopes.c
    29 46.4743535762 lines/sec
    Ran 3 tests, found 0 unexpected mismatches